docs: add proper license and attribution comments for Hono and pathpida references
