+++ src/helper/server/types.ts
@@ -1,7 +1,4 @@
-// Inspired by Hono (https://github.com/honojs/hono)
-// Some parts of this code are based on or adapted from the Hono project
-
-import { HttpMethod } from "../../lib/types";
+import type { ValidationSchema } from "./route-types";
 import type { NextResponse, NextRequest } from "next/server";
 
 type KnownContentType =
@@ -375,27 +372,7 @@ export interface RouteContext<
   ) => TypedNextResponse<undefined, TStatus, "">;
 }
 
-export type RouteResponse =
-  | TypedNextResponse
-  | Promise<TypedNextResponse | void>;
-
-export type RequiredRouteResponse =
-  | TypedNextResponse
-  | Promise<TypedNextResponse>;
-
-export interface RouteBindings {
-  params?: Params | Promise<Params>;
-  query?: Query;
-}
-
 export type ValidationTarget = "params" | "query";
-
-export interface ValidationSchema {
-  // eslint-disable-next-line @typescript-eslint/no-empty-object-type
-  input: {};
-  // eslint-disable-next-line @typescript-eslint/no-empty-object-type
-  output: {};
-}
 
 type ValidationFor<
   TDirection extends keyof ValidationSchema,
@@ -423,88 +400,3 @@ export type ConditionalValidationInput<
 > = TTarget extends TExpected
   ? ValidationInputFor<TTarget, TSchema>
   : TFallback;
-
-export type Handler<
-  TParams = Params,
-  TQuery = Query,
-  TValidationSchema extends ValidationSchema = ValidationSchema,
-  TRouteResponse extends RouteResponse = RouteResponse,
-> = (
-  routeContext: RouteContext<TParams, TQuery, TValidationSchema>
-) => TRouteResponse;
-
-export type ErrorHandler<
-  TRouteResponse extends RequiredRouteResponse,
-  TParams = Params,
-  TQuery = Query,
-  TValidationSchema extends ValidationSchema = ValidationSchema,
-> = (
-  error: unknown,
-  routeContext: RouteContext<TParams, TQuery, TValidationSchema>
-) => TRouteResponse;
-
-type RouteHandler<
-  TParams extends RouteBindings["params"],
-  TRouteResponse extends RouteResponse,
-> = (
-  req: NextRequest,
-  segmentData: { params: Promise<TParams> }
-  // Exclude void | undefined because a response is always returned or an error is thrown internally
-) => Promise<Exclude<Awaited<TRouteResponse>, void | undefined>>;
-
-type HttpMethodMapping<
-  THttpMethod extends HttpMethod,
-  TParams extends RouteBindings["params"],
-  TRouteResponse extends RouteResponse,
-> = Record<THttpMethod, RouteHandler<TParams, TRouteResponse>>;
-
-export interface MethodRouteDefinition<
-  THttpMethod extends HttpMethod,
-  TBindings extends RouteBindings,
-  TOnErrorResponse extends RequiredRouteResponse,
-  TParams extends TBindings["params"] = TBindings extends {
-    params: infer TValue;
-  }
-    ? Awaited<TValue>
-    : Params,
-  TQuery extends TBindings["query"] = TBindings extends { query: infer TValue }
-    ? TValue
-    : Query,
-> {
-  // 1 handler
-  <
-    TV1 extends ValidationSchema = ValidationSchema,
-    TR1 extends RequiredRouteResponse = RequiredRouteResponse,
-  >(
-    handler: Handler<TParams, TQuery, TV1, TR1>
-  ): HttpMethodMapping<THttpMethod, TParams, TR1 | TOnErrorResponse>;
-
-  // 2 handlers
-  <
-    TV1 extends ValidationSchema = ValidationSchema,
-    TV2 extends ValidationSchema = TV1,
-    TR1 extends RouteResponse = RouteResponse,
-    TR2 extends RequiredRouteResponse = RequiredRouteResponse,
-  >(
-    handler1: Handler<TParams, TQuery, TV1, TR1>,
-    handler2: Handler<TParams, TQuery, TV2, TR2>
-  ): HttpMethodMapping<THttpMethod, TParams, TR1 | TR2 | TOnErrorResponse>;
-
-  // 3 handlers
-  <
-    TV1 extends ValidationSchema = ValidationSchema,
-    TV2 extends ValidationSchema = TV1,
-    TV3 extends ValidationSchema = TV1 & TV2,
-    TR1 extends RouteResponse = RouteResponse,
-    TR2 extends RouteResponse = RouteResponse,
-    TR3 extends RequiredRouteResponse = RequiredRouteResponse,
-  >(
-    handler1: Handler<TParams, TQuery, TV1, TR1>,
-    handler2: Handler<TParams, TQuery, TV2, TR2>,
-    handler3: Handler<TParams, TQuery, TV3, TR3>
-  ): HttpMethodMapping<
-    THttpMethod,
-    TParams,
-    TR1 | TR2 | TR3 | TOnErrorResponse
-  >;
-}
